Add USDUSB support for tests.

+++ usdusb/Manager.cc
@@ -46,7 +46,7 @@ namespace Pds {
 	  ;
 	}
 	else {
-	  long nRecords = MAX_RECORDS;
+	  int nRecords = MAX_RECORDS;
           const int tmo = 100; // milliseconds
 	  int status = USB4_ReadFIFOBufferStruct(_dev, 
 						 &nRecords,
